Signature, tracing, endpoint fix
The sha1 signature did not work for me. Replaced with SHA256.

Also making sure the endpoint ends with '/' -- without it the base
string for the signature does not match the original Amazon assumes and
signature verification fails.

Replaced GET with POST. Not sure why GET does not work, but it does not
for me.

Added exception catching to the example for Exception::Class newbies
like myself.